requirements.txt
# Vibecast Requirements
# MAKER-powered entertainment discovery

# Core dependencies
asyncio-compat>=0.2.0

# LLM Providers (install as needed)
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Optional: API server
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
